Implement symmetric storage encryption with AES-GCM
The values of the Irmin storage (that is, not the key IDs) are now
encrypted with AES in GCM mode.  The symmetric key is for now
provided with the --masterkey command line argument, which takes a
16, 24 or 32 byte key in hex format.  The container image start
script supports a KEYFENDER_MASTERKEY environment variable, that sets
this commandline accordingly.  If no masterkey is provided, the
storage is still in plaintext as before and a warning is printed.